chore: reorganize global package structure
