+++ src/main/java/com/synergy/backend/global/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.synergy.backend.global.security;
+package com.synergy.backend.global.jwt;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -13,6 +13,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.synergy.backend.domain.member.entity.RoleType;
 import com.synergy.backend.global.common.ApiResponse;
+import com.synergy.backend.global.security.CustomUserDetailsService;
 import com.synergy.backend.global.security.exception.UnKnownUserTypeException;
 
 import jakarta.servlet.FilterChain;
